+++ client/model_update_node_pools_amazon.go
@@ -12,7 +12,7 @@ package client
 
 type UpdateNodePoolsAmazon struct {
 	Autoscaling bool `json:"autoscaling,omitempty"`
-	Count float32 `json:"count,omitempty"`
-	MinCount float32 `json:"minCount,omitempty"`
-	MaxCount float32 `json:"maxCount,omitempty"`
+	Count int32 `json:"count,omitempty"`
+	MinCount int32 `json:"minCount,omitempty"`
+	MaxCount int32 `json:"maxCount,omitempty"`
 }
